Consolodate RIF home health claims into service periods instead of one claim per encounter (typically one per day).
